+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change Log
 
 ### [Unreleased][unreleased]
+
+### [1.1.0] - 2024-02-02
 ####  Added
 - Added publish, subscribe commands
 - Added rename command
@@ -83,7 +85,8 @@
 #### Added
 - Wrapper for \Redis
 
-[unreleased]: https://github.com/lulco/redis-proxy/compare/1.0.0...HEAD
+[unreleased]: https://github.com/lulco/redis-proxy/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/lulco/redis-proxy/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/lulco/redis-proxy/compare/0.7.1...1.0.0
 [0.7.1]: https://github.com/lulco/redis-proxy/compare/0.7.0...0.7.1
 [0.7.0]: https://github.com/lulco/redis-proxy/compare/0.6.0...0.7.0